+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ### Fixed
 
 - Class statement lookup used for fields now finds definitions before declarations.
+- Exception when the user wrote infinite `include` statements.
 
 ## [0.10.0] - 2025-08-21