+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## [Unreleased]
 
+## [0.11.0] - 2025-09-01
+
 ### Added
 
 - Line markers to `let` body items that navigates to the previous `let` statement that is overridden.
@@ -167,7 +169,8 @@ items in the `Find usage` functionality confusing for users.
 
 - Support for using `tblgen-lsp-server` built as part of the opened CMake project
 
-[Unreleased]: https://github.com/zero9178/IntelliJ-MLIR-ODS/compare/v0.10.0...HEAD
+[Unreleased]: https://github.com/zero9178/IntelliJ-MLIR-ODS/compare/v0.11.0...HEAD
+[0.11.0]: https://github.com/zero9178/IntelliJ-MLIR-ODS/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/zero9178/IntelliJ-MLIR-ODS/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/zero9178/IntelliJ-MLIR-ODS/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/zero9178/IntelliJ-MLIR-ODS/compare/v0.7.0...v0.8.0